+++ common/src/main/java/com/microsoft/identity/common/internal/cache/HelloCache.kt
@@ -28,7 +28,6 @@ import android.os.Build
 import androidx.annotation.VisibleForTesting
 import com.microsoft.identity.common.java.interfaces.INameValueStorage
 import com.microsoft.identity.common.java.interfaces.IPlatformComponents
-import com.microsoft.identity.common.java.util.StringUtil
 import com.microsoft.identity.common.logging.Logger
 import java.util.concurrent.TimeUnit
 
@@ -49,25 +48,28 @@ import java.util.concurrent.TimeUnit
  * @param protocolName         name of the protocol that invokes hello().
  * @param targetAppPackageName package name of the app that this client will hello() with.
  * @param components           Platform components.
- * @param timeoutInMs         Cache entry timeout.
+ * @param cacheExpiryTimeInMs  Cache entry expiry time.
  */
 open class HelloCache (
     private val context: Context,
     private val protocolName: String,
     private val targetAppPackageName: String,
     components: IPlatformComponents,
-    private val timeoutInMs: Long = sTimeoutInMs
+    private val cacheExpiryTimeInMs: Long = DEFAULT_CACHE_EXPIRY_MILLIS
 ) {
     companion object {
         private val TAG = HelloCache::class.java.simpleName
         private const val SHARED_PREFERENCE_NAME = "com.microsoft.common.ipc.hello.cache"
         private var sIsEnabled = true
-        private val sTimeoutInMs = TimeUnit.HOURS.toMillis(4)
+
+        /**
+         * Default life time of cache entry.
+         */
+        private val DEFAULT_CACHE_EXPIRY_MILLIS = TimeUnit.HOURS.toMillis(4)
 
         /**
          * If set to false, Hello cache will be disabled.
          * When you're developing protocol change, you might not want the cache to be enabled.
-         *
          *
          * For debugging only.
          */
@@ -114,11 +116,11 @@ open class HelloCache (
             Logger.error(methodTag, "Failed to retrieve key", e)
             return null
         }
-        val negotiationValue = fileManager[key]
-        if (StringUtil.isNullOrEmpty(negotiationValue)) {
+        val cachedRawValue = fileManager[key]
+        if (cachedRawValue.isNullOrEmpty()) {
             return null
         }
-        val cacheResult = HelloCacheResult.deserialize(negotiationValue!!)
+        val cacheResult = HelloCacheResult.deserialize(cachedRawValue)
         if (cacheResult == null) {
             Logger.info(methodTag, "Legacy or invalid cache value.")
             fileManager.remove(key)
@@ -126,7 +128,7 @@ open class HelloCache (
         }
 
         // check if expired. Delete entry and return null.
-        if ((System.currentTimeMillis() - cacheResult.timeStamp) > timeoutInMs) {
+        if ((System.currentTimeMillis() - cacheResult.timeStamp) > cacheExpiryTimeInMs) {
             Logger.info(methodTag, "Cache entry is expired.")
             fileManager.remove(key)
             return null
@@ -236,117 +238,4 @@ open class HelloCache (
                 packageInfo.versionCode.toString()
             }
         }
-
-    /**
-     * Internal helper class to store and retrieve cache value in safe manner.
-     * @param negotiatedProtocolVersion Stores either negotiated protocol value or handshake error.
-     * @param timeStamp Time stamp of entry.
-     */
-    class HelloCacheResult (
-        val negotiatedProtocolVersion: String?,
-        val error: String?,
-        internal val timeStamp: Long
-    ) {
-        init {
-            if (!(negotiatedProtocolVersion.isNullOrEmpty() xor error.isNullOrEmpty())) {
-                throw IllegalStateException("Either both parameters provided or none provide.")
-            }
-        }
-
-        companion object {
-            private val TAG = HelloCacheResult::class.java.simpleName
-            private const val SEPARATOR = ","
-            private const val ERROR_PREFIX = "E"
-            private const val SUCCESS_PREFIX = "S"
-            private const val SUCCESS_CACHE_VALUE_FORMAT = "$SUCCESS_PREFIX$SEPARATOR%s$SEPARATOR%d"
-            private const val ERROR_CACHE_VALUE_FORMAT = "$ERROR_PREFIX$SEPARATOR%s$SEPARATOR%d"
-            private const val HANDSHAKE_ERROR = "handshake_error"
-
-            /**
-             * Reads raw entry from cache into HelloCacheValue object.
-             * @param value cache entry read from file expected to be in format <negotiated protocol version>,<timestamp>
-             * @return null if value is not valid.
-            </timestamp></negotiated> */
-            internal fun deserialize(value: String): HelloCacheResult? {
-                val methodTag = "$TAG:deserialize"
-                val values = value.split(SEPARATOR)
-                // valid value is <String>,<TimeStamp>
-                if (values.size != 3) {
-                    Logger.warn(methodTag, "Legacy or Invalid cache entry. $value")
-                    return null
-                }
-
-                return try {
-                    val timeStamp = values[2].toLong()
-                    if (values[0] == ERROR_PREFIX) {
-                        createError(values[1], timeStamp)
-                    } else {
-                        createFromNegotiatedProtocolVersion(values[1], timeStamp)
-                    }
-                } catch (e: NumberFormatException) {
-                    Logger.error(methodTag, "Invalid cache entry. $value", e)
-                    null
-                }
-            }
-
-            /**
-             * Constructs new successful HelloCacheResult using timeStamp as current time.
-             * @param negotiatedProtocolVersion A valid negotiated protocol version value e.g. 13.0, 14.0
-             */
-            internal fun createFromNegotiatedProtocolVersion(negotiatedProtocolVersion: String): HelloCacheResult {
-                return createFromNegotiatedProtocolVersion(negotiatedProtocolVersion, System.currentTimeMillis())
-            }
-
-            /**
-             * Constructs new HelloCacheResult with handshake error using timeStamp as current time.
-             */
-            internal fun createHandshakeError(): HelloCacheResult {
-                return createError(HANDSHAKE_ERROR, System.currentTimeMillis())
-            }
-
-            private fun createFromNegotiatedProtocolVersion(negotiatedProtocolVersion: String, timeStamp: Long): HelloCacheResult {
-                return HelloCacheResult(negotiatedProtocolVersion, null, timeStamp)
-            }
-
-            private fun createError(error: String, timeStamp: Long): HelloCacheResult {
-                return HelloCacheResult(null, error, timeStamp)
-            }
-        }
-
-        /**
-         * Returns true if entry contains handshake_failure.
-         */
-        fun isHandShakeError() : Boolean {
-            return isError() && error == HANDSHAKE_ERROR
-        }
-
-        /**
-         * Returns true if the result contains negotiated protocol version value.
-         */
-        fun isSuccess() : Boolean {
-            return !negotiatedProtocolVersion.isNullOrEmpty()
-        }
-
-        /**
-         * Returns true if entry contains an error.
-         */
-        fun isError() : Boolean {
-            return !error.isNullOrEmpty()
-        }
-
-        /**
-         * Value will be serialized as below
-         * For successful negotiated protocol value (NBP): S,<NBP>,<timestamp> e.g S,13.0,16000700
-         * For caching any error: S,<error>,<timestamp> e.g S,handshake_failure,16000700
-         * This code should work for all error and success in general but currently the only
-         * error supported is handshake_failure.
-         */
-        internal fun serialize(): String {
-            return if (!error.isNullOrEmpty()) {
-                String.format(ERROR_CACHE_VALUE_FORMAT, error, timeStamp)
-            } else {
-                String.format(SUCCESS_CACHE_VALUE_FORMAT, negotiatedProtocolVersion, timeStamp)
-            }
-        }
-    }
 }
